Add read by batch
